GWTII-182: tms plugin: refactor packages
