+++ plugin/tms/example-jar/src/main/java/org/geomajas/gwt2/plugin/tms/example/client/sample/TmsLayerPanel.ui.xml
