--- plugin/tms/example-jar/src/main/java/org/geomajas/plugin/tms/example/client/i18n/SampleMessages.java
@@ -1,34 +1,0 @@
-/*
- * This is part of Geomajas, a GIS framework, http://www.geomajas.org/.
- *
- * Copyright 2008-2014 Geosparc nv, http://www.geosparc.com/, Belgium.
- *
- * The program is available in open source according to the GNU Affero
- * General Public License. All contributions in this program are covered
- * by the Geomajas Contributors License Agreement. For full licensing
- * details, see LICENSE.txt in the project root.
- */
-
-package org.geomajas.plugin.tms.example.client.i18n;
-
-import com.google.gwt.i18n.client.Messages;
-
-/**
- * Specific messages for the many samples.
- *
- * @author Pieter De Graef
- */
-public interface SampleMessages extends Messages {
-
-	String tmsLayerTitle();
-
-	String tmsLayerShort();
-
-	String tmsLayerDescription();
-
-	String capabilitiesTitle();
-
-	String capabilitiesShort();
-
-	String capabilitiesDescription();
-}
