+++ plugin/tms/example-jar/src/main/resources/org/geomajas/gwt2/plugin/tms/example/ExampleJar.gwt.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+
+<!--
+  ~ This is part of Geomajas, a GIS framework, http://www.geomajas.org/.
+  ~
+  ~ Copyright 2008-2014 Geosparc nv, http://www.geosparc.com/, Belgium.
+  ~
+  ~ The program is available in open source according to the GNU Affero
+  ~ General Public License. All contributions in this program are covered
+  ~ by the Geomajas Contributors License Agreement. For full licensing
+  ~ details, see LICENSE.txt in the project root.
+  -->
+
+<module rename-to="examplejar">
+	<inherits name='org.geomajas.gwt2.example.base.ExampleBase' />
+	<inherits name='org.geomajas.gwt2.plugin.tms.GeomajasClientPluginTms' />
+    <entry-point class="org.geomajas.gwt2.plugin.tms.example.client.ExampleJar"/>
+	<source path='client'/>
+</module>
