+++ plugin/tms/example-jar/src/main/resources/org/geomajas/gwt2/plugin/tms/example/client/i18n/SampleMessages.properties
