+++ plugin/tms/tms/src/main/resources/org/geomajas/gwt2/plugin/tms/GeomajasClientPluginTms.gwt.xml
